Added filtering,sorting,pagination to get tasks of user

+++ src/routers/task.js
@@ -1,11 +1,15 @@
 const Task = require('../models/task');
-const express = require('express')
+const express = require('express');
+const auth = require('../middleware/auth');
 const router = new express.Router();
 
 
-router.post('/tasks', async (req, res) => {
-
-    const task = new Task(req.body);
+router.post('/tasks', auth, async (req, res) => {
+    // const task = new Task(req.body);
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    })
     try {
         await task.save();
         res.status(201).send(task);
@@ -20,7 +24,52 @@ router.post('/tasks', async (req, res) => {
     //     res.status(400).send(err)
     // });
 });
-router.patch('/tasks/:id', async (req, res) => {
+
+//Get /tasks?completed=true //Filteration
+//Get /tasks?limit=10&skip=20 //Pagination
+//Get /tasks?sortBy=createdAt:desc
+
+router.get('/tasks', auth, async (req, res) => {
+    const match = {};
+    const sort = {}
+
+    if (req.query.completed) {
+        match.completed = req.query.completed === 'true'
+    }
+    if (req.query.sortBy) {
+        const parts = req.query.sortBy.split(':');
+        sort[parts[0]] = parts[1] === 'desc' ? -1 : 1
+    }
+
+    try {
+        await req.user.populate({
+            path: 'tasks',
+            match,
+            options: {
+                limit: parseInt(req.query.limit),
+                skip: parseInt(req.query.skip),
+                sort
+            }
+        }).execPopulate();
+        res.send(req.user.tasks)
+    } catch (err) {
+        res.status(500).send()
+    }
+})
+
+router.get('/tasks/:id', auth, async (req, res) => {
+    const _id = req.params.id;
+    try {
+        const task = await Task.findOne({ _id, owner: req.user._id });
+        if (!task) {
+            return res.status(404).send();
+        }
+        res.send(task);
+    } catch (err) {
+        res.status(500).send(err)
+    }
+})
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body);
     const allowdUpdates = ['description', 'completed'];
     const isValidOperation = updates.every((update) => allowdUpdates.includes(update))
@@ -32,22 +81,27 @@ router.patch('/tasks/:id', async (req, res) => {
     try {
         const _id = req.params.id;
 
-        const task = await Task.findById(_id);
-        updates.forEach(update => { task[update] = req.body[update] })
-        await task.save()
-        // const result = await Task.findByIdAndUpdate(_id, req.body, { new: true, runValidators: true });
+        // const task = await Task.findById(_id);
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id })
+
         if (!task) {
             return res.status(404).send("Failed")
         }
+        updates.forEach(update => { task[update] = req.body[update] })
+        await task.save()
+        // const result = await Task.findByIdAndUpdate(_id, req.body, { new: true, runValidators: true });
+
         res.send(task)
     } catch (err) {
         res.status(500).send(err);
     }
 
 });
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     try {
-        const user = await Task.findByIdAndDelete(req.params.id);
+        // const user = await Task.findByIdAndDelete(req.params.id);
+        const user = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id })
+
         if (!user) {
             return res.status(404).send()
         }
